Split parse into multiple files
